Modify data sent in from AddEthanol Form
