+++ src/functions/calculateCocktailData.ts
@@ -4,21 +4,7 @@ import { getMacros } from './getMacros';
 import { getDilutionIngredients } from './getDilutionIngredients';
 import { dilutionCalculus } from './dilutionCalculus';
 import { Technique } from '../types/Technique';
-
-
-export interface CocktailOutput {
-  name: string;
-  ethanol: number;
-  carbs: number;
-  calories: number;
-  fat: number;
-  protein: number;
-  sugar: number;
-  addedSugar: number;
-  ounces: number;
-  abv: number;
-  sugarAcid: number;
-}
+import { CocktailOutput } from '../types/CocktailOutput';
 
 export function calculateCocktailData(cocktail: Spec[], spiritsData: SpiritData[], technique: Technique): CocktailOutput {
