+++ src/hooks/useManageDrinks.ts
@@ -3,9 +3,10 @@ import { useDrinks } from '../context/DrinksContext';
 import { toast } from 'react-toastify';
 import { Spec } from '../types/Spec';
 import { SpiritData } from '../types/SpiritData';
-import { getMacros } from '../functions/getMacros';
 import { useAuth } from '../context/AuthContext';
 import { DrinksCocktail } from '../types/DrinksCocktail';
+import { calculateCocktailData } from '../functions/calculateCocktailData';
+import { calculateDrinkData } from '../functions/calculateDrinkData';
 
 interface ManageDrinksOptions {
     checkAuth?: boolean;
@@ -18,21 +19,6 @@ interface ManageDrinksOptions {
 export const useManageDrinks = (spiritData?: SpiritData[]) => {
     const { setDrinks, setCocktail } = useDrinks();
     const { auth } = useAuth();
-
-    
-    const createDrink = (data) => ({
-        name: data.name || 'Cocktail from Nutrition/Dilution', 
-        ethanol: data.ethanol || 0,
-        carbs: data.carbs || 0, 
-        calories: data.calories || 0,
-        fat: data.fat || 0,
-        protein: data.protein || 0,
-        sugar: data.sugar || 0,
-        addedsugar: data.addedsugar || 0,
-        ounces: data.ounces || 0,
-        abv: data.abv || 0,
-        createdAt: new Date(),
-    });
 
     const addDrinkToDB = useCallback(async (drink: DrinksCocktail) => {
         try {
@@ -68,14 +54,18 @@ export const useManageDrinks = (spiritData?: SpiritData[]) => {
         }
 
         if (info && info.ounces > 0 && info.abv > 0) {
-            const drinksCocktail = createDrink(info);
-            setDrinks((prevDrinks) => [...prevDrinks, drinksCocktail]);
-            addDrinkToDB(drinksCocktail);
+            //From Add Ethanol Form
+            const cocktailData = calculateDrinkData(info);
+            setDrinks((prevDrinks) => [...prevDrinks, cocktailData]);
+            addDrinkToDB(cocktailData);
         } else if (specs && specs.length > 0) {
-            const drinkDetails = getMacros(specs, spiritData);
-            const drinksCocktail = createDrink(drinkDetails);
-            setDrinks((prevDrinks) => [...prevDrinks, drinksCocktail]);
-            addDrinkToDB(drinksCocktail);
+            //From Liquid Form and Cocktail PageCalculus(getDilutionIngredients(specs, spiritData), "shaken");
+            console.log(specs)
+            const cocktailData = calculateCocktailData(specs, spiritData, "shaken");
+
+            console.log(cocktailData)
+            //setDrinks((prevDrinks) => [...prevDrinks, cocktailData]);
+            //addDrinkToDB(cocktailData);
         } else {
             toast.error("You cannot add an empty drink");
             return;
