+++ src/types/CocktailOutput.ts
@@ -1,0 +1,13 @@
+export interface CocktailOutput {
+    name: string;
+    ethanol: number;
+    carbs: number;
+    calories: number;
+    fat: number;
+    protein: number;
+    sugar: number;
+    addedSugar: number;
+    ounces: number;
+    abv: number;
+    sugarAcid: number;
+  }
